requirements.txt
# Core application dependencies
streamlit>=1.36.0
pydantic-ai-slim[groq]>=0.6.0
pydantic>=2.0.0
pydantic-settings>=2.0.0

# Data analysis dependencies
pandas>=2.0.0
numpy>=1.24.0
matplotlib>=3.7.0
seaborn>=0.12.0
jupyter-client>=8.0.0

# Training dependencies (optional - only needed for model training)
datasets>=2.14.0
transformers>=4.48.0
trl>=0.7.0
peft>=0.6.0

# Testing dependencies (optional - only needed for running tests)
pytest>=8.3.0
